Add new controller
